test_requirements.txt
pytest
pytest-django
pytest-pythonpath
mock

factory_boy